% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/categorical.R
\name{analyze.categorical}
\alias{analyze.categorical}
\title{Analyzing categorical dependent variables}
\usage{
analyze.categorical(theframe, theindependentfactortoken,
  thedependentfactortoken)
}
\arguments{
\item{theframe}{the frame to which the factors belong}

\item{theindependentfactortoken}{the index of the independent variable}

\item{thedependentfactortoken}{the index of the dependent variable}
}
\value{
the p value of the test (either Fisher or Kruskal).
}
\description{
\code{analyze.categorical} provides a p-value for a test of correlation
between a categorical dependent variable and either a categorical or ordinal
independent variable.
}

