% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ordinal.R
\name{analyze.ordinal}
\alias{analyze.ordinal}
\title{Analyzing ordinal dependent variables}
\usage{
analyze.ordinal(theframe, theindependentfactortoken, thedependentfactortoken)
}
\arguments{
\item{theframe}{the frame to which the factors belong}

\item{theindependentfactortoken}{the index of the independent variable}

\item{thedependentfactortoken}{the index of the dependent variable}
}
\value{
the p value of the test (either Kruskal or Kendall).
}
\description{
\code{analyze.ordinal} provides a p-value for a test of correlation
between an ordinal dependent variable and either a categorical or ordinal
independent variable.
}

