% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{shred}
\alias{shred}
\title{Destroying tokens}
\usage{
shred(token)
}
\arguments{
\item{token}{the token to be destroyed}
}
\value{
nothing
}
\description{
\code{shred} completely destroys the specified token.
}

