% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/categorical.R
\name{summarize.categorical}
\alias{summarize.categorical}
\title{Summarizing categorical responses}
\usage{
summarize.categorical(theframe, thefactortoken)
}
\arguments{
\item{theframe}{the frame to which the factor belongs}

\item{thefactortoken}{the index of the factor}
}
\value{
a labelled list of summary results:
\description{
 \item{dependent}{the question being summarized}
 \item{N}{the number of nontrivial responses}
 \item{mode}{the most populous response}
}
}
\description{
\code{summarize.categorical} provides summary statistics for a categorical
variable.
}

