% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ordinal.R
\name{summarize.ordinal}
\alias{summarize.ordinal}
\title{Summarizing ordinal responses}
\usage{
summarize.ordinal(theframe, thefactortoken)
}
\arguments{
\item{theframe}{the frame to which the factor belongs}

\item{thefactortoken}{the index of the factor}
}
\value{
a labelled list of summary results:
\description{
 \item{dependent}{the question being summarized}
 \item{N}{the number of nontrivial responses}
 \item{midquartiles}{the lower, median, and upper quartiles}
}
}
\description{
\code{summarize.ordinal} provides summary statistics for an ordinal
variable.
}

