% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarizeAll}
\alias{summarizeAll}
\title{Summarizing multiple factors}
\usage{
summarizeAll(theframe, thetokenvector)
}
\arguments{
\item{theframe}{the data frame to which the indices belong}

\item{thetokenvector}{the vector of indices to be summarized}
}
\value{
a vector of lists containing summary statistics
}
\description{
\code{summarizeAll} takes a data frame and a vector of indices and produces summary statistics
for all indices.
}

