% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generic.R
\name{theNOf}
\alias{theNOf}
\title{Counting nontrivial responses}
\usage{
theNOf(theframe, thefactortoken)
}
\arguments{
\item{theframe}{the frame to which a question belongs}

\item{thefactortoken}{the index of the factor}
}
\value{
the number of nontrivial questions
}
\description{
\code{theNOf} counts the number of nontrivial responses to a question.
}

